Module Map.Multi

type nonrec 'a t = 'a list t

val rev_union : 'a t -> 'a t -> 'a t

val cons : 'a t -> key -> 'a -> 'a t

val find : 'a t -> key -> 'a list

val add_all : 'a t -> key -> 'a list -> 'a t

val find_elt : 'a t -> f:('a -> bool) -> (key * 'a) option

find_elt m ~f linearly traverses the map m and the contained lists to find the first element e (in a list l, mapped to key k) such that f e = true. If such an e is found then the function returns Some (k,e), otherwise it returns None.

val to_flat_list : 'a t -> 'a list

val equal : 'a t -> 'a t -> equal:('a -> 'a -> bool) -> bool

val map : 'a t -> f:('a -> 'b) -> 'b t

val to_dyn : ('a -> Dyn.t) -> 'a t -> Dyn.t